Changes index.js structure
